requirements.txt
# Core dependencies
torch>=2.1.0
torchvision>=0.16.0
timm>=1.0.3
numpy>=1.26.0
pandas>=2.1.0
matplotlib>=3.8.0
scikit-learn>=1.3.0
scipy>=1.11.0
seaborn>=0.13.0
opencv-python>=4.9.0
Pillow>=10.0.0
streamlit>=1.33
# Optional - for notebooks & training utilities
tqdm>=4.66.0
jupyter>=1.0.0
ipykernel>=6.25.0

# XAI (if you use Grad-CAM, SHAP, etc.)
grad-cam>=1.5.2
shap>=0.44.1

# System utilities
psutil>=5.9.0
requests>=2.31.0
# Web app (optional)
streamlit>=1.28.0
gradio>=3.50.0